+++ ConsoleWoW-master/.gitignore
