+++ ConsoleWoW-master/src/no/hiof/sondrli/consolewow/Priest.java
